Switch to linenoise instead of readline

+++ t/clitests.sh
@@ -21,16 +21,22 @@ if [ -z "${1:-}" ]; then
     echo "$0 <binary>"
     exit 1
 fi
+tater=$1
+
+set +feuo pipefail
+echo -e "let a = 1;\nprint a; exit(77);" | ${tater} --debug --gc-trace --gc-stress
+if [ ${?} != 77 ]; then
+    exit 1
+fi
+set -feuo pipefail
+
+${tater} /tmp/nosuchfileordir || true # expected failure
+${tater} /tmp/nosuchfileordir noarg || true # expected failure
 
 TEST_TMPDIR=$(mktemp -d)
 trap "rm -rf ${TEST_TMPDIR};" err exit
-
-tater=$1
-echo 'let a = 1;' | ${tater}
-${tater} /tmp/nosuchfileordir || true # expected failure
-${tater} /tmp/nosuchfileordir noarg || true # expected failure
-echo 'let a = 1;' > "${TEST_TMPDIR}/t.tot"
-${tater} "${TEST_TMPDIR}/t.tot"
+echo -e "let a = 1;\nprint a;" > "${TEST_TMPDIR}/t.tot"
+${tater} --debug --gc-trace --gc-stress "${TEST_TMPDIR}/t.tot"
 ${tater} --version
 ${tater} --help
 ${tater} --debug --gc-trace --gc-stress --invalid-option-will-fail || true
